Improve downloader script
